Id Card Image Reader

+++ ImageReader/Models/UsageDto.cs
@@ -2,5 +2,5 @@ namespace ImageReader.Models;
 public class UsageDto
 {
     public long PromptTokens { get; set; }
-    public int? TotalTokens { get; set; }
+    public long? TotalTokens { get; set; }
 }
